refactor (deploy): configuration and gems for new server deploy on EC2 (#105)
* refactor (deploy): reinstalls capistrano for new server

* refactor (deploy): adds capistrano rvm

* fix (deploy): changes deploy directory to home directory

* refactor (deploy): appends application yml file

* refactor (deploy): installed nginx

* refactor (deploy): testing in new server

* refactor (deploy): more puma tasks

* refactor (deploy): more puma tasks

* refactor (deploy): puma deploy new

* fix (deploy): removes puma.rb form linked files

* refactor (deploy): changed directory to glownet

* fix (deploy): adds puma.rb

* fix (deploy): more puma

* fix (deploy): ammend to last commit

* fix (nginx): added propper config

* refactor (server): removes puma and uses passenger

* fix (ruby): changed to ruby 2.3.3

* fix (ransack): search_field instead of input for forms

* feature (git): adds eb ignores

* refactor (deploy): saves copy of sidekiq and nginx config

+++ db/migrate/20161222130559_unite_settings_in_event.rb
@@ -1,46 +1,46 @@
 class UniteSettingsInEvent < ActiveRecord::Migration
   def change
-    add_column :events, :city_mandatory, :boolean
-    add_column :events, :phone_mandatory, :boolean
-    add_column :events, :gender_mandatory, :boolean
-    add_column :events, :address_mandatory, :boolean
-    add_column :events, :country_mandatory, :boolean
-    add_column :events, :postcode_mandatory, :boolean
-    add_column :events, :birthdate_mandatory, :boolean
-    add_column :events, :agreed_event_condition, :boolean
-    add_column :events, :receive_communications, :boolean
-    add_column :events, :receive_communications_two, :boolean
+    add_column :events, :city_mandatory, :boolean unless column_exists?(:events, :city_mandatory, :boolean)
+    add_column :events, :phone_mandatory, :boolean unless column_exists?(:events, :phone_mandatory, :boolean)
+    add_column :events, :gender_mandatory, :boolean unless column_exists?(:events, :gender_mandatory, :boolean)
+    add_column :events, :address_mandatory, :boolean unless column_exists?(:events, :address_mandatory, :boolean)
+    add_column :events, :country_mandatory, :boolean unless column_exists?(:events, :country_mandatory, :boolean)
+    add_column :events, :postcode_mandatory, :boolean unless column_exists?(:events, :postcode_mandatory, :boolean)
+    add_column :events, :birthdate_mandatory, :boolean unless column_exists?(:events, :birthdate_mandatory, :boolean)
+    add_column :events, :agreed_event_condition, :boolean unless column_exists?(:events, :agreed_event_condition, :boolean)
+    add_column :events, :receive_communications, :boolean unless column_exists?(:events, :receive_communications, :boolean)
+    add_column :events, :receive_communications_two, :boolean unless column_exists?(:events, :receive_communications_two, :boolean)
 
-    add_column :events, :gtag_format, :string, default: "standard"
-    add_column :events, :gtag_type, :string, default: 'ultralight_c'
-    add_column :events, :gtag_deposit, :integer, default: 0
-    add_column :events, :maximum_gtag_balance, :integer, default: 300
-    add_column :events, :cards_can_refund, :boolean, default: true
-    add_column :events, :wristbands_can_refund, :boolean, default: true
+    add_column :events, :gtag_format, :string, default: "standard" unless column_exists?(:events, :gtag_format, :string)
+    add_column :events, :gtag_type, :string, default: 'ultralight_c' unless column_exists?(:events, :gtag_type, :string)
+    add_column :events, :gtag_deposit, :integer, default: 0 unless column_exists?(:events, :gtag_deposit, :integer)
+    add_column :events, :maximum_gtag_balance, :integer, default: 300 unless column_exists?(:events, :maximum_gtag_balance, :integer)
+    add_column :events, :cards_can_refund, :boolean, default: true unless column_exists?(:events, :cards_can_refund, :boolean)
+    add_column :events, :wristbands_can_refund, :boolean, default: true unless column_exists?(:events, :wristbands_can_refund, :boolean)
 
-    add_column :events, :min_apk_version, :string, default: "1.0"
-    add_column :events, :uid_reverse, :boolean, default: false
-    add_column :events, :topup_initialize_gtag, :boolean, default: true
-    add_column :events, :pos_update_online_orders, :boolean, default: false
-    add_column :events, :touchpoint_update_online_orders, :boolean, default: false
-    add_column :events, :cypher_enabled, :boolean, default: false
-    add_column :events, :fast_removal_password, :string, default: '123456'
-    add_column :events, :private_zone_password, :string, default: '123456'
+    add_column :events, :min_apk_version, :string, default: "1.0" unless column_exists?(:events, :min_apk_version, :string)
+    add_column :events, :uid_reverse, :boolean, default: false unless column_exists?(:events, :uid_reverse, :boolean)
+    add_column :events, :topup_initialize_gtag, :boolean, default: true unless column_exists?(:events, :topup_initialize_gtag, :boolean)
+    add_column :events, :pos_update_online_orders, :boolean, default: false unless column_exists?(:events, :pos_update_online_orders, :boolean)
+    add_column :events, :touchpoint_update_online_orders, :boolean, default: false unless column_exists?(:events, :touchpoint_update_online_orders, :boolean)
+    add_column :events, :cypher_enabled, :boolean, default: false unless column_exists?(:events, :cypher_enabled, :boolean)
+    add_column :events, :fast_removal_password, :string, default: '123456' unless column_exists?(:events, :fast_removal_password, :string)
+    add_column :events, :private_zone_password, :string, default: '123456' unless column_exists?(:events, :private_zone_password, :string)
 
-    add_column :events, :sync_time_gtags, :integer, default: 10
-    add_column :events, :sync_time_tickets, :integer, default: 5
-    add_column :events, :transaction_buffer, :integer, default: 100
-    add_column :events, :days_to_keep_backup, :integer, default: 5
-    add_column :events, :sync_time_customers, :integer, default: 10
-    add_column :events, :sync_time_server_date, :integer, default: 1
-    add_column :events, :sync_time_basic_download, :integer, default: 5
-    add_column :events, :sync_time_event_parameters, :integer, default: 1
+    add_column :events, :sync_time_gtags, :integer, default: 10 unless column_exists?(:events, :sync_time_gtags, :integer)
+    add_column :events, :sync_time_tickets, :integer, default: 5 unless column_exists?(:events, :sync_time_tickets, :integer)
+    add_column :events, :transaction_buffer, :integer, default: 100 unless column_exists?(:events, :transaction_buffer, :integer)
+    add_column :events, :days_to_keep_backup, :integer, default: 5 unless column_exists?(:events, :days_to_keep_backup, :integer)
+    add_column :events, :sync_time_customers, :integer, default: 10 unless column_exists?(:events, :sync_time_customers, :integer)
+    add_column :events, :sync_time_server_date, :integer, default: 1 unless column_exists?(:events, :sync_time_server_date, :integer)
+    add_column :events, :sync_time_basic_download, :integer, default: 5 unless column_exists?(:events, :sync_time_basic_download, :integer)
+    add_column :events, :sync_time_event_parameters, :integer, default: 1 unless column_exists?(:events, :sync_time_event_parameters, :integer)
 
-    add_column :events, :ultralight_c_private_key, :string
-    add_column :events, :mifare_classic_public_key, :string
-    add_column :events, :mifare_classic_private_key_a, :string
-    add_column :events, :mifare_classic_private_key_b, :string
-    add_column :events, :ultralight_ev1_private_key, :string
+    add_column :events, :ultralight_c_private_key, :string unless column_exists?(:events, :ultralight_c_private_key, :string )
+    add_column :events, :mifare_classic_public_key, :string unless column_exists?(:events, :mifare_classic_public_key, :string )
+    add_column :events, :mifare_classic_private_key_a, :string unless column_exists?(:events, :mifare_classic_private_key_a, :string )
+    add_column :events, :mifare_classic_private_key_b, :string unless column_exists?(:events, :mifare_classic_private_key_b, :string )
+    add_column :events, :ultralight_ev1_private_key, :string unless column_exists?(:events, :ultralight_ev1_private_key, :string )
 
     Event.all.each do |event|
       event.city_mandatory = event.registration_settings["city"]
@@ -83,14 +83,14 @@ class UniteSettingsInEvent < ActiveRecord::Migration
       event.save(validate: false)
     end
 
-    remove_column :events, :registration_settings, :jsonb
-    remove_column :events, :gtag_settings, :jsonb
-    remove_column :events, :device_settings, :jsonb
-    remove_column :events, :min_apk_version, :string
-    remove_column :events, :cypher_enabled, :boolean
-    remove_column :events, :host_country, :string
-    remove_column :events, :location, :string
-    remove_column :events, :url, :string
+    remove_column :events, :registration_settings, :jsonb if column_exists?(:events, :registration_settings)
+    remove_column :events, :gtag_settings, :jsonb if column_exists?(:events, :gtag_settings)
+    remove_column :events, :device_settings, :jsonb if column_exists?(:events, :device_settings)
+    remove_column :events, :min_apk_version, :string if column_exists?(:events, :min_apk_version)
+    remove_column :events, :cypher_enabled, :boolean if column_exists?(:events, :cypher_enabled)
+    remove_column :events, :host_country, :string if column_exists?(:events, :host_country)
+    remove_column :events, :location, :string if column_exists?(:events, :location)
+    remove_column :events, :url, :string if column_exists?(:events, :url)
   end
 end
