+++ db/migrate/20161223162409_remove_description_from_tickets.rb
@@ -1,5 +1,5 @@
 class RemoveDescriptionFromTickets < ActiveRecord::Migration
   def change
-    remove_column :tickets, :description, :string
+    remove_column :tickets, :description, :string if column_exists?(:tickets, :description)
   end
 end
