+++ db/migrate/20170103145554_remove_gtag_name_from_event_translations.rb
@@ -1,5 +1,5 @@
 class RemoveGtagNameFromEventTranslations < ActiveRecord::Migration
   def change
-    remove_column :event_translations, :gtag_name, :string
+    remove_column :event_translations, :gtag_name, :string if column_exists?(:event_translations, :gtag_name)
   end
 end
